requirements-ml.txt
# Core AI and ML libraries
langchain>=0.2.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
sentence-transformers>=5.1.1
transformers>=4.56.2
torch>=2.8.0
torch-geometric>=2.6.1

# Vector databases and embeddings
chromadb>=1.1.0
pinecone-client==2.2.4
weaviate-client==3.25.3

# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.11.9
python-multipart==0.0.6

# Database and storage
neo4j>=5.18.0
redis>=5.0.1

# Data processing and NLP
numpy>=2.3.3
pandas>=2.3.3
scikit-learn>=1.7.2
nltk>=3.8.1
spacy>=3.7.2
jsonlines>=4.0.0

# Graph processing and ML
networkx>=3.5
pyvis>=0.3.2
plotly>=5.17.0
scipy>=1.16.2

# BioPython for scientific data
biopython>=1.83

# Monitoring and observability
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
structlog>=23.2.0

# Utilities
python-dotenv>=1.1.1
httpx>=0.28.1
aiofiles>=23.2.1
pydantic-settings>=2.1.0
tqdm>=4.67.1

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

